+++ src/main/java/com/practice/hrbank/repository/EmployeeRepository.java
@@ -6,13 +6,18 @@ import java.time.Instant;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 public interface EmployeeRepository extends JpaRepository<Employee, Long>,
-    JpaSpecificationExecutor<Employee>, EmployeeRepositoryCustom {
+    JpaSpecificationExecutor<Employee> {
+
+  Page<Employee> findAll(Specification<Employee> spec, Pageable pageable);
 
   @Query("SELECT e.employeeNumber FROM Employee e WHERE e.employeeNumber LIKE :yearPrefix ORDER BY e.employeeNumber DESC")
   String findLatestEmployeeNumberByYear(@Param("yearPrefix") int yearPrefix);
